move login to components
